+++ Sources/SystemNotification/SystemNotificationEdge.swift
@@ -8,18 +8,16 @@
 
 import SwiftUI
 
-/**
- This enum specifies the edges from which a notification can
- currently be presented.
- */
+/// This enum defines edges from which a system notification
+/// can be presented.
 public enum SystemNotificationEdge {
     
     case top, bottom
     
     public var alignment: Alignment {
         switch self {
-        case .top: return .top
-        case .bottom: return .bottom
+        case .top: .top
+        case .bottom: .bottom
         }
     }
 }