+++ Sources/SystemNotification/SystemNotificationMessageStyle.swift
@@ -8,37 +8,35 @@
 
 import SwiftUI
 
-/**
- This style can be used with ``SystemNotificationMessage``s.
- 
- You can easily create new styles like this:
- 
- ```swift
- extension SystemNotificationMessageStyle {
- 
-    static var custom = Self(iconColor: .yellow)
- }
- ```
-
- The ``standard`` style will be used by default if you don't
- provide a custom style.
- */
+/// This style can style a ``SystemNotificationMessage``.
+///
+/// Apply ``SwiftUI/View/systemNotificationMessageStyle(_:)``
+/// if an app or notification should use a custom style.
+///
+/// You can easily create a custom style like this:
+///
+/// ```swift
+/// extension SystemNotificationMessageStyle {
+///    static var custom = Self(iconColor: .yellow)
+/// }
+/// ```
+///
+/// The ``standard`` value is used by default when you don't
+/// apply a custom value.
 public struct SystemNotificationMessageStyle {
 
-    /**
-     Create a system notification message style.
-
-     - Parameters:
-       - iconColor: The color to apply to the icon.
-       - iconFont: The font to apply to the icon.
-       - iconTextSpacing: The spacing to apply between the icon and the text.
-       - padding: The padding to add to the content.
-       - textColor: The color to apply to the text.
-       - textFont: The font to apply to the text.
-       - titleColor: The color to apply to the title.
-       - titleFont: The font to apply to the title.
-       - titleTextSpacing: The spacing to apply between the title and the text.
-     */
+    /// Create a custom system notification message style.
+    ///
+    /// - Parameters:
+    ///   - iconColor: The color to apply to the icon.
+    ///   - iconFont: The font to apply to the icon.
+    ///   - iconTextSpacing: The spacing to apply between the icon and the text.
+    ///   - padding: The padding to add to the content.
+    ///   - textColor: The color to apply to the text.
+    ///   - textFont: The font to apply to the text.
+    ///   - titleColor: The color to apply to the title.
+    ///   - titleFont: The font to apply to the title.
+    ///   - titleTextSpacing: The spacing to apply between the title and the text.
     public init(
         iconColor: Color = .primary.opacity(0.6),
         iconFont: Font = Font.title3,
@@ -60,14 +58,6 @@ public struct SystemNotificationMessageStyle {
         self.titleFont = titleFont
         self.titleTextSpacing = titleTextSpacing
     }
-    
-    /**
-     This is a standard notification message configuration.
-
-     You can apply a custom value to this property to affect
-     all system notification within an app.     
-     */
-    public static var standard = SystemNotificationMessageStyle()
     
     /// The color to apply to the icon.
     public var iconColor: Color
@@ -95,4 +85,38 @@ public struct SystemNotificationMessageStyle {
     
     /// The spacing to apply between the title and the text.
     public var titleTextSpacing: CGFloat
+}
+
+public extension SystemNotificationMessageStyle {
+
+    /// The standard system notification message style.
+    ///
+    /// You can set this style to change the global default.
+    static var standard = Self()
+}
+
+public extension View {
+
+    /// Apply a ``SystemNotificationMessageStyle`` to the view.
+    func systemNotificationMessageStyle(
+        _ style: SystemNotificationMessageStyle
+    ) -> some View {
+        self.environment(\.systemNotificationMessageStyle, style)
+    }
+}
+
+private extension SystemNotificationMessageStyle {
+
+    struct Key: EnvironmentKey {
+
+        static var defaultValue: SystemNotificationMessageStyle = .standard
+    }
+}
+
+public extension EnvironmentValues {
+
+    var systemNotificationMessageStyle: SystemNotificationMessageStyle {
+        get { self [SystemNotificationMessageStyle.Key.self] }
+        set { self [SystemNotificationMessageStyle.Key.self] = newValue }
+    }
 }